Fix blank video when switching on/off screen
This happens when using `ExoPlayer.setVideoEffects()`.

This CL also fixes the first frame not rendered problem, originally solved in
https://github.com/androidx/media/commit/038408dd99491b93added7c7c8a9bb51bb6981ac, but rolled back in https://github.com/androidx/media/commit/3836a361a301f2f0fe6716a02bef08dbfce85ece because the solution introduces
the flash that is observed in b/292111083.

Before media3 1.1 release, the output size of `VideoFrameProcessor` is not
reported to the app. This was changed later after introducing
`CompositingVideoSinkProvider`, where the video size after processing **is**
reported to the app. After this CL, the size is again, not reported.

PiperOrigin-RevId: 602345087
(cherry picked from commit fb3155ea3768d9685a91b802ed30febd602e415e)