Test: Fix failing HdrEditingTest by swapping assumption.
This was broken by http://https://github.com/androidx/media/commit/88342530ddc6f9d88aec5a2d8e0181249d7d61f0, which added an assumes
method, but missed the `!` operator for the tests where HDR is
unsupported. Add an assumes method that assumes a lack of support
for HDR.

Tested on the failing device (Pixel 7) and confirmed this fixes
the test on that device, to throw AssumptionViolatedException
instead of allowing test logic to run after failing the
assumption.

PiperOrigin-RevId: 616846588